Split handler/action/tests.rs into per-category modules
